+++ src/cc/lib/swcdb/core/FileUtils.cc
@@ -147,7 +147,7 @@ bool mkdirs(const std::string& dirname) {
   *(tmpdir+dirname.length()) = '/';
 
   int saved_errno = 0;
-  for(int n=1; n < dirname.length()+1; ++n){
+  for(size_t n=1; n < dirname.length()+1; ++n){
     if(*(tmpdir+n) != '/')
       continue;
     *(tmpdir+n) = 0;
